requirements.txt
transformers>=4.42
sentencepiece
timm
ftfy
safetensors
pillow
torch
torchvision
accelerate
tqdm
pandas
requests
bitsandbytes
